+++ ext/tiny_tds/result.c
@@ -6,10 +6,10 @@
 
 VALUE cTinyTdsResult;
 extern VALUE mTinyTds, cTinyTdsClient, cTinyTdsError;
-VALUE cBigDecimal, cDate;
+VALUE cKernel, cDate;
 VALUE opt_decimal_zero, opt_float_zero, opt_one, opt_zero, opt_four, opt_19hdr, opt_onek, opt_tenk, opt_onemil, opt_onebil;
 static ID intern_new, intern_utc, intern_local, intern_localtime, intern_merge,
-          intern_civil, intern_new_offset, intern_plus, intern_divide;
+          intern_civil, intern_new_offset, intern_plus, intern_divide, intern_bigd;
 static ID sym_symbolize_keys, sym_as, sym_array, sym_cache_rows, sym_first, sym_timezone, sym_local, sym_utc, sym_empty_sets;
 
 
@@ -228,7 +228,7 @@ static VALUE rb_tinytds_result_fetch_row(VALUE self, ID timezone, int symbolize_
           int data_slength = (int)data_info->precision + (int)data_info->scale + 1;
           char converted_decimal[data_slength];
           dbconvert(rwrap->client, coltype, data, data_len, SYBVARCHAR, (BYTE *)converted_decimal, -1);
-          val = rb_funcall(cBigDecimal, intern_new, 1, rb_str_new2((char *)converted_decimal));
+          val = rb_funcall(cKernel, intern_bigd, 1, rb_str_new2((char *)converted_decimal));
           break;
         }
         case SYBFLT8: {
@@ -246,7 +246,7 @@ static VALUE rb_tinytds_result_fetch_row(VALUE self, ID timezone, int symbolize_
           char converted_money[25];
           long long money_value = ((long long)money->mnyhigh << 32) | money->mnylow;
           sprintf(converted_money, "%" LONG_LONG_FORMAT, money_value);
-          val = rb_funcall(cBigDecimal, intern_new, 2, rb_str_new2(converted_money), opt_four);
+          val = rb_funcall(cKernel, intern_bigd, 2, rb_str_new2(converted_money), opt_four);
           val = rb_funcall(val, intern_divide, 1, opt_tenk);
           break;
         }
@@ -254,7 +254,7 @@ static VALUE rb_tinytds_result_fetch_row(VALUE self, ID timezone, int symbolize_
           DBMONEY4 *money = (DBMONEY4 *)data;
           char converted_money[20];
           sprintf(converted_money, "%f", money->mny4 / 10000.0);
-          val = rb_funcall(cBigDecimal, intern_new, 1, rb_str_new2(converted_money));
+          val = rb_funcall(cKernel, intern_bigd, 1, rb_str_new2(converted_money));
           break;
         }
         case SYBBINARY:
@@ -566,7 +566,7 @@ static VALUE rb_tinytds_result_insert(VALUE self) {
 
 void init_tinytds_result() {
   /* Data Classes */
-  cBigDecimal = rb_const_get(rb_cObject, rb_intern("BigDecimal"));
+  cKernel = rb_const_get(rb_cObject, rb_intern("Kernel"));
   cDate = rb_const_get(rb_cObject, rb_intern("Date"));
   /* Define TinyTds::Result */
   cTinyTdsResult = rb_define_class_under(mTinyTds, "Result", rb_cObject);
@@ -588,6 +588,7 @@ void init_tinytds_result() {
   intern_new_offset = rb_intern("new_offset");
   intern_plus = rb_intern("+");
   intern_divide = rb_intern("/");
+  intern_bigd = rb_intern("BigDecimal");
   /* Symbol Helpers */
   sym_symbolize_keys = ID2SYM(rb_intern("symbolize_keys"));
   sym_as = ID2SYM(rb_intern("as"));